move plugin to legacy
