+++ go.work
@@ -4,5 +4,5 @@ use (
 	.
 	./examples/apiserver
 	./logging
-	./plugin
+	./plugin-legacy
 )
