+++ plugin-legacy/errors.go
