+++ plugin-legacy/kubeconfig/initializer.go
