+++ plugin-legacy/kubeconfig/initializer_test.go
@@ -6,7 +6,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"k8s.io/client-go/rest"
 
-	"github.com/grafana/grafana-app-sdk/plugin/kubeconfig"
+	"github.com/grafana/grafana-app-sdk/plugin-legacy/kubeconfig"
 )
 
 func TestCachingInitializer(t *testing.T) {
