+++ plugin-legacy/kubeconfig/middleware.go
@@ -3,11 +3,10 @@ package kubeconfig
 import (
 	"context"
 
+	"github.com/grafana/grafana-app-sdk/plugin-legacy"
+	"github.com/grafana/grafana-app-sdk/plugin-legacy/router"
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/log"
-
-	"github.com/grafana/grafana-app-sdk/plugin"
-	"github.com/grafana/grafana-app-sdk/plugin/router"
 )
 
 // LoadingMiddleware returns a new middleware that can be used on a router for
