+++ plugin-legacy/router/vars.go
