+++ plugin-legacy/router/vars_test.go
